bug fix, fix score summary styling
